pyproject.toml
[project]
name = "llm-assesment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.0",
    "fastapi>=0.120.4",
    "langchain>=1.0.3",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.2",
    "langchain-openai>=1.0.1",
    "langchain-text-splitters>=1.0.0",
    "pypdf>=6.1.3",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",
    "requests>=2.32.0",
    "python-dotenv>=1.2.1",
]
